+++ src/main/java/cz/venglarjakub/ppj/pictures/repository/MongoPictureRepository.java
@@ -5,10 +5,12 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 /**
  * Created by Jakub Venglar on 25. 4. 2016.
  */
 @Profile("prod-mongo")
 @Repository
-public interface MongoPictureRepository extends MongoRepository<Picture, Integer>, PictureRepository {
+public interface MongoPictureRepository extends MongoRepository<Picture, BigInteger>, PictureRepository {
 }
